requirements.txt
Flask==2.3.2
google-cloud-datastore==2.18.0